backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
ollama==0.1.7
pandas==2.1.4
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1